+++ Doctrine/CouchDB/UserListener.php
@@ -11,13 +11,34 @@
 
 namespace FOS\UserBundle\Doctrine\CouchDB;
 
+use Doctrine\Common\EventSubscriber;
 use Doctrine\ODM\CouchDB\Event;
 use Doctrine\ODM\CouchDB\Event\LifecycleEventArgs;
 use FOS\UserBundle\Model\UserInterface;
-use FOS\UserBundle\Doctrine\AbstractUserListener;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
-class UserListener extends AbstractUserListener
+class UserListener implements EventSubscriber
 {
+    /**
+     * @var \FOS\UserBundle\Model\UserManagerInterface
+     */
+    private $userManager;
+
+    /**
+     * @var ContainerInterface
+     */
+    private $container;
+
+    /**
+     * Constructor
+     *
+     * @param ContainerInterface $container
+     */
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+    }
+
     public function getSubscribedEvents()
     {
         return array(
@@ -29,7 +50,7 @@ class UserListener extends AbstractUserListener
     /**
      * @param LifecycleEventArgs $args
      */
-    public function prePersist($args)
+    public function prePersist(LifecycleEventArgs $args)
     {
         $object = $args->getDocument();
         if ($object instanceof UserInterface) {
@@ -40,11 +61,26 @@ class UserListener extends AbstractUserListener
     /**
      * @param LifecycleEventArgs $args
      */
-    public function preUpdate($args)
+    public function preUpdate(LifecycleEventArgs $args)
     {
         $object = $args->getDocument();
         if ($object instanceof UserInterface) {
             $this->updateUserFields($object);
         }
+    }
+
+    /**
+     * Updates the user properties.
+     *
+     * @param UserInterface $user
+     */
+    private function updateUserFields(UserInterface $user)
+    {
+        if (null === $this->userManager) {
+            $this->userManager = $this->container->get('fos_user.user_manager');
+        }
+
+        $this->userManager->updateCanonicalFields($user);
+        $this->userManager->updatePassword($user);
     }
 }